Move executables, dlls and winflexbison data to new bin directory
This is a preparation for being able to install perl (which expects a
lib directory in a directory above where the perl exeutable is) while
not having to specify more than one path to the executables in this
repo. Also it's a cleaner solution.

winflexbison expects to find a data directory in the same directory
where bison.exe is. Although this can be overridden with
BISON_PKGDATADIR, moving the data directory to bin avoids the hassle
of setting it in various MSBuild project files or CMake configuration
files. Also to the winflexbison custom_build_rules directory has been
moved although it seems unused.